# docs/swagger.yaml
basePath: /
definitions:
  internal_controllers.AddToCartInput:
    properties:
      product_id:
        example: 1
        type: integer
      quantity:
        example: 2
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
  internal_controllers.CartItem:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      product:
        $ref: '#/definitions/internal_controllers.Product'
      product_id:
        example: 1
        type: integer
      quantity:
        example: 2
        type: integer
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  internal_controllers.CartResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_controllers.CartItem'
        type: array
      total:
        example: 1999.98
        type: number
    type: object
  internal_controllers.CheckoutResponse:
    properties:
      message:
        example: order placed
        type: string
      order_id:
        example: 1
        type: integer
      total:
        example: 1999.98
        type: number
    type: object
...
  internal_controllers.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  internal_controllers.MessageResponse:
    properties:
      message:
        example: success
        type: string
    type: object
...
  internal_controllers.Order:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      order_items:
        items:
          $ref: '#/definitions/internal_controllers.OrderItem'
        type: array
      status:
        example: pending
        type: string
      total_price:
        example: 1999.98
        type: number
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  internal_controllers.OrderItem:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      order_id:
        example: 1
        type: integer
      price:
        example: 999.99
        type: number
      product:
        $ref: '#/definitions/internal_controllers.Product'
      product_id:
        example: 1
        type: integer
      quantity:
        example: 2
        type: integer
      subtotal:
        example: 1999.98
        type: number
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
  internal_controllers.PaymentIntentResponse:
    properties:
      amount:
        example: 1999.98
        type: number
      payment_intent:
        example: 1
        type: integer
      redirect_url:
        example: http://localhost:8080/pay-gateway?intent=1
        type: string
    type: object
  internal_controllers.PaymentWebhookResponse:
    properties:
      message:
        example: payment successful
        type: string
      order:
        example: 1
        type: integer
    type: object
  internal_controllers.Product:
    properties:
      category:
        example: Electronics
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        example: High-performance laptop
        type: string
      id:
        example: 1
        type: integer
      image_url:
        example: https://example.com/image.jpg
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Laptop
        type: string
      price:
        example: 999.99
        type: number
      quantity:
        example: 10
        type: integer
      slug:
        example: laptop
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
  internal_controllers.UpdateCartQuantityInput:
    properties:
      quantity:
        example: 3
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ecommerce.com
    name: API Support
  description: This is a backend service for the Go e-commerce project with comprehensive
    authentication, product management, cart, orders, and payment functionality.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-Commerce API
  version: "1.0"
paths:
...
  /api/cart:
    delete:
      description: Removes all items from the user's cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Clear entire cart
      tags:
      - Cart
    get:
      description: Retrieves the user's shopping cart with all items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.CartResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's cart
      tags:
      - Cart
  /api/cart/{id}:
    delete:
      description: Removes a specific item from the cart
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Updates the quantity of a specific cart item
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: string
      - description: New Quantity
        in: body
        name: quantity
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.UpdateCartQuantityInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - Cart
  /api/cart/add:
    post:
      consumes:
      - application/json
      description: Adds a product to the user's shopping cart
      parameters:
      - description: Cart Item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.AddToCartInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.MessageResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /api/orders/:
    get:
      description: Retrieves all orders for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_controllers.Order'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's orders
      tags:
      - Orders
  /api/orders/{id}:
    get:
      description: Retrieves details of a specific order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get order details
      tags:
      - Orders
  /api/orders/checkout:
    post:
      description: Creates an order from the user's cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.CheckoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Checkout cart
      tags:
      - Orders
  /api/payment/start/{orderId}:
    post:
      description: Initiates payment for an order
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.PaymentIntentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start payment process
      tags:
      - Payment
  /api/payment/webhook:
    post:
      description: Handles payment gateway callbacks
      parameters:
      - description: Payment Intent ID
        in: query
        name: intent_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.PaymentWebhookResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      summary: Payment webhook
      tags:
      - Payment
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"